+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Deprecated the `BaseSource.integration` property in favor of `TypedSource.type`.
 
+## Added
+
+- Added support for `THEOlive` events.
+
 ## [9.7.1]
 
 ### Fixed